initializing single outfit component
