+++ client/src/components/SingleOutfit/SingleOutfit.js
@@ -1,0 +1,14 @@
+import React from "react";
+
+
+function outfit() {
+    return (
+        <>
+       
+
+        
+        </>
+    )
+};
+
+export default outfit;
