+++ controllers/outfitsController.js
@@ -1,0 +1,18 @@
+const db = require("../models");
+
+// Defining methods for the outfitsController
+module.exports = {
+  findAll: function(req, res) {
+    db.Outfits
+      .find(req.query)
+      .sort({ date: -1 })
+      .then(dbModel => res.json(dbModel))
+      .catch(err => res.status(422).json(err));
+  },
+  findById: function(req, res) {
+    db.Outfits
+      .findById(req.params.id)
+      .then(dbModel => res.json(dbModel))
+      .catch(err => res.status(422).json(err));
+  }
+}
